+++ app/src/main/java/org/hyperledger/besu/controller/BesuController.java
@@ -369,7 +369,8 @@ public class BesuController implements java.io.Closeable {
         }
         builder = new CliqueBesuControllerBuilder();
       } else {
-        throw new IllegalArgumentException("Unknown consensus mechanism defined");
+        LOG.warn("No consensus mechanism detected in genesis config, using PoS");
+        return new MergeBesuControllerBuilder().genesisConfig(genesisConfig);
       }
 
       // wrap with TransitionBesuControllerBuilder if we have a terminal total difficulty: